Fixing compilation with glibc 2.28+ and bison 3.0.4 (included in v18.06.1).
This commit adds a patch to bison 3.0.4, which is still included in
OpenWrt 18.06.1 (the 18.06 HEAD instead has a newer version of bison),
which is in turn the stable version on which this V2X platform is based.
The patch is by "rdslw" and it has been directly imported from:
https://github.com/openwrt/openwrt/pull/1290
It is needed in order to compile OpenWrt-V2X on newer versions of
Ubuntu, such as Ubuntu 19.04.