Update APU2 configuration files and add enhanced uqmi files
This commit adds to "configs" two additional files, containing
more up-to-date .config files for the APU2 boards.
It also adds inside the /root directory of the final image a few
files related to uqmi (used for managing LTE/cellular connections
with OpenWrt-V2X), that can be used under an x86_64 architecture
in case any issue is experienced with the connection to a
cellular network.